1. Add save scale size methods.

+++ app/src/main/java/com/rulhouse/ruler/feature_node/domain/model/Measurement.kt
@@ -6,8 +6,8 @@ import androidx.room.PrimaryKey
 @Entity
 data class Measurement(
     val title: String,
-    val width: Int,
-    val height: Int,
+    val width: Float,
+    val height: Float,
     val timeStamp: Long,
     @PrimaryKey val id: Int? = null
 ) {
